pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "twitch-alerts"
requires-python = ">=3.10"
description = "Alert me when streamers are live"
readme = "README.md"
license = "MIT"
authors = [
    { email = "preocts@preocts.com", name = "Preocts" },
]
maintainers = []
keywords = []
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: Implementation :: CPython",
]
dynamic = ["version"]
dependencies = [
    "httpx",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
format = [
    "black",
    "isort",
]
lint = [
    "flake8",
    "flake8-builtins",
    "flake8-pep585",
    "mypy",
]
dev = [
    "pytest",
    "pytest-randomly",
    "coverage",
    "vcrpy",
]

[project.urls]
homepage = "https://github.com/Preocts/twitch-alerts"

# [project.scripts]
# python-src-example = "twitch_alerts.sample:main"

[tool.hatch.version]
source = "vcs"

[tool.black]
line-length = 100

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_incomplete_defs = false
disallow_untyped_defs = false
warn_unused_ignores = false

[tool.coverage.run]
branch = true
source = ["tests"]
source_pkgs = ["twitch_alerts"]

[tool.coverage.paths]
source = ["src/", "*/site-packages", "*/src"]
test = ["tests/", "*/tests"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "\\.\\.\\.",
    "if TYPE_CHECKING:",
]
